core/src/agent/mod.rs
//! ML*/Agent Checker System
//! 
//! This module implements the core ML*/Agent Checker system, providing a comprehensive framework
//! for managing and coordinating AI/ML agents in a distributed environment. The system
//! handles various aspects including agent lifecycle management, performance monitoring,
//! security enforcement, and ML model integration.
//!
//! # Architecture
//!
//! The ML*/Agent Checker system is built around several key components:
//!
//! - Agent Management: Core agent lifecycle and coordination
//! - Performance Analysis: Real-time performance monitoring and optimization
//! - Security Framework: Comprehensive security controls and validation
//! - ML Integration: Seamless integration with ML models and services
//! - Metrics Collection: Detailed metrics gathering and analysis
//! - Validation System: Robust validation of system components
//!
//! # Features
//!
//! - Comprehensive agent lifecycle management
//! - Real-time performance monitoring and optimization
//! - Security enforcement and validation
//! - ML model integration and coordination
//! - Metrics collection and analysis
//! - System validation and health checks
//! - Testing and debugging support
//!
//! # Example
//!
//! ```rust
//! use anya::agent::{Agent, AgentConfig};
//!
//! async fn start_agent() -> Result<(), AgentError> {
//!     // Create agent configuration
//!     let config = AgentConfig::new()
//!         .with_name("ml_agent_1")
//!         .with_security_level(SecurityLevel::High)
//!         .with_performance_monitoring(true);
//!
//!     // Create and start agent
//!     let agent = Agent::new(config);
//!     agent.start().await?;
//!
//!     Ok(())
//! }
//! ```

mod error;
mod testing;
mod security;
mod metrics;
mod ml_integration;

pub use error::AgentError;
pub use testing::{TestManager, TestStatus};
pub use security::{SecurityValidator, SecurityStatus, ComplianceStatus};
pub use metrics::{MetricsCollector, ComponentMetrics};
pub use ml_integration::MLIntegration;

/// ML*/Agent Checker System
/// Manages component lifecycle, testing, and deployment stages
pub struct MLAgentChecker {
    system_manager: Arc<SystemManager>,
    ml_service: Arc<MLService>,
    security_manager: Arc<dyn SecurityManager>,
    metrics: Arc<RwLock<UnifiedMetrics>>,
    state: Arc<RwLock<AgentState>>,
    test_manager: Arc<TestManager>,
    security_validator: Arc<SecurityValidator>,
    metrics_collector: Arc<MetricsCollector>,
    ml_integration: Arc<MLIntegration>,
}

/// Agent State tracking
#[derive(Debug, Clone)]
pub struct AgentState {
    /// Component completion status (0-100%)
    completion_status: HashMap<String, f64>,
    /// Current stage for each component
    stage_status: HashMap<String, StageStatus>,
    /// Test results
    test_results: HashMap<String, TestStatus>,
    /// Security validation status
    security_status: HashMap<String, SecurityStatus>,
}

/// Stage status for components
#[derive(Debug, Clone, PartialEq)]
pub enum StageStatus {
    Development,  // 60% threshold
    Production,   // 90% threshold
    Release,      // 99% threshold
}

impl MLAgentChecker {
    /// Create new ML*/Agent Checker instance
    pub fn new(
        system_manager: Arc<SystemManager>,
        ml_service: Arc<MLService>,
        security_manager: Arc<dyn SecurityManager>,
        metrics: Arc<RwLock<UnifiedMetrics>>,
    ) -> Self {
        let test_manager = Arc::new(TestManager::new());
        let security_validator = Arc::new(SecurityValidator::new(security_manager.clone()));
        let metrics_collector = Arc::new(MetricsCollector::new(metrics.clone()));
        let ml_integration = Arc::new(MLIntegration::new(
            ml_service.clone(),
            system_manager.get_model_repository(),
            metrics.clone(),
            security_manager.clone(),
        ));

        Self {
            system_manager,
            ml_service,
            security_manager,
            metrics,
            state: Arc::new(RwLock::new(AgentState::default())),
            test_manager,
            security_validator,
            metrics_collector,
            ml_integration,
        }
    }

    /// Check component readiness and trigger appropriate actions
    pub async fn check_component_readiness(&self, component: &SystemComponent) -> Result<StageStatus, AgentError> {
        // Get ML-based readiness analysis
        let ml_readiness = self.ml_integration.analyze_component_readiness(component).await?;
        
        // Get test status
        let test_status = self.test_manager.get_test_status(component).await?;
        
        // Get security status
        let security_status = self.security_validator.get_security_status(component).await?;
        
        // Get metrics
        let metrics = self.metrics_collector.get_component_metrics(component).await?;
        
        // Calculate weighted completion
        let completion = self.calculate_completion_percentage(
            ml_readiness,
            &test_status,
            &security_status,
            &metrics,
        ).await?;
        
        let mut state = self.state.write().await;
        
        // Update completion status
        state.completion_status.insert(component.name.clone(), completion);
        
        // Determine stage based on completion
        let stage = match completion {
            x if x >= 99.0 => StageStatus::Release,
            x if x >= 90.0 => StageStatus::Production,
            x if x >= 60.0 => StageStatus::Development,
            _ => return Err(AgentError::NotReady),
        };
        
        // Update stage status
        state.stage_status.insert(component.name.clone(), stage.clone());
        
        // Trigger appropriate actions based on stage
        self.trigger_stage_actions(component, &stage).await?;
        
        Ok(stage)
    }

    /// Calculate weighted completion percentage
    async fn calculate_completion_percentage(
        &self,
        ml_readiness: f64,
        test_status: &TestStatus,
        security_status: &SecurityStatus,
        metrics: &ComponentMetrics,
    ) -> Result<f64, AgentError> {
        // Weights for different factors
        const ML_WEIGHT: f64 = 0.3;
        const TEST_WEIGHT: f64 = 0.25;
        const SECURITY_WEIGHT: f64 = 0.25;
        const METRICS_WEIGHT: f64 = 0.2;
        
        // Calculate test score
        let test_score = if test_status.unit_tests_passed && test_status.integration_tests_passed {
            test_status.coverage_percentage
        } else {
            0.0
        };
        
        // Calculate security score
        let security_score = if security_status.validation_passed && security_status.audit_completed {
            100.0
        } else {
            0.0
        };
        
        // Calculate metrics score
        let metrics_score = (metrics.performance_score + 
                           metrics.reliability_score + 
                           metrics.security_score) / 3.0;
        
        // Calculate weighted completion
        let completion = (ml_readiness * ML_WEIGHT) +
                        (test_score * TEST_WEIGHT) +
                        (security_score * SECURITY_WEIGHT) +
                        (metrics_score * METRICS_WEIGHT);
                        
        Ok(completion)
    }

    /// Trigger appropriate actions based on stage
    async fn trigger_stage_actions(&self, component: &SystemComponent, stage: &StageStatus) -> Result<(), AgentError> {
        match stage {
            StageStatus::Development => {
                // Run development tests
                self.test_manager.run_development_tests(component).await?;
                // Run development security validation
                self.security_validator.run_development_validation(component).await?;
                // Collect development metrics
                self.metrics_collector.collect_development_metrics(component).await?;
            },
            StageStatus::Production => {
                // Run production tests
                self.test_manager.run_production_tests(component).await?;
                // Run production security validation
                self.security_validator.run_production_validation(component).await?;
                // Collect production metrics
                self.metrics_collector.collect_production_metrics(component).await?;
            },
            StageStatus::Release => {
                // Run release tests
                self.test_manager.run_release_tests(component).await?;
                // Run release security validation
                self.security_validator.run_release_validation(component).await?;
                // Collect release metrics
                self.metrics_collector.collect_release_metrics(component).await?;
            }
        }
        
        Ok(())
    }
}

impl Default for AgentState {
    fn default() -> Self {
        Self {
            completion_status: HashMap::new(),
            stage_status: HashMap::new(),
            test_results: HashMap::new(),
            security_status: HashMap::new(),
        }
    }
}

#[cfg(test)]
mod tests {
    use super::*;
    use crate::ml::service::MockMLService;
    use crate::security::MockSecurityManager;
    
    #[tokio::test]
    async fn test_component_readiness() {
        // Test implementation
    }
    
    #[tokio::test]
    async fn test_stage_transitions() {
        // Test implementation
    }
    
    #[tokio::test]
    async fn test_ml_integration() {
        // Test implementation
    }
}
